# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WAGE - API BI",
        "version": "0.1"
    },
    "paths": {
...
        "/api/vendas-classe": {
            "get": {
                "tags": [
                    "VendasClasse"
                ],
                "summary": "Retorna a lista de registro de vendas por classe",
                "description": "Retorna a lista de registro de vendas por classe, usando como parametro de pesquisa o cnpj_empresa / codigo_classe / ano / mes / dia ",
                "operationId": "totVendClassIndex",
                "parameters": [
                    {
                        "name": "cnpj_empresa",
                        "in": "query",
                        "description": "Cnpj da empresa ao qual o registro pertence - orbigatorio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "codigo_classe",
                        "in": "query",
                        "description": "Codigo da classe ao qual o registro pertence",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ano",
                        "in": "query",
                        "description": "Usado para pesquisa quando somente este campos esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mes",
                        "in": "query",
                        "description": "Usado para pesquisa quando o campo |Ano| tambem esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dia",
                        "in": "query",
                        "description": "Usado para pesquisa quando somente este campos esta preenchido",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a lista de registros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendasClasseWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "VendasClasse"
                ],
                "summary": "Insere o registro no sistema",
                "description": "",
                "operationId": "totVendClassStore",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/VendasClasse"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna registro inserido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendasClasseWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                }
            }
        },
        "/api/vendas-classe/{id}": {
            "get": {
                "tags": [
                    "VendasClasse"
                ],
                "summary": "Retorna o registro",
                "description": "Retorna o registro, usando como parametro o id",
                "operationId": "totVendClassShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o registro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendasClasseWithDate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "login nao autorizado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "VendasClasse"
                ],
                "summary": "Edita o registro",
                "description": "Edita o registro selecionado, usando como parametro o id",
                "operationId": "totVendClassUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/VendasClasse"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o registro alterado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VendasClasseWithDate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro nao encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "VendasClasse"
                ],
                "summary": "Remove o registro",
                "description": "Remove o registro selecionado, usando como parametro o id",
                "operationId": "totVendClassDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro excluido com sucesso"
                    },
                    "404": {
                        "description": "Registro nao encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VendasClasseWithDate": {
                "properties": {
                    "id": {
                        "description": "ID do registro",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "data": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "mes": {
                        "description": "",
                        "type": "integer",
                        "example": "12",
                        "nullable": false
                    },
                    "ano": {
                        "description": "",
                        "type": "integer",
                        "example": "2025",
                        "nullable": false
                    },
                    "cnpj_empresa": {
                        "description": "",
                        "type": "string",
                        "example": "39311444000183",
                        "nullable": false
                    },
                    "codigo_classe": {
                        "description": "",
                        "type": "integer",
                        "example": "",
                        "nullable": false
                    },
                    "nome_classe": {
                        "description": "",
                        "type": "string",
                        "example": "",
                        "nullable": false
                    },
                    "vendas_quant": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "vendas_bruto": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "vendas_desconto": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "vendas_liquido": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "itens_por_venda": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "itens_dif_venda": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "itens_quant_vendas": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "devolucoes_quant": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "devolucoes_valor": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "devolucoes_taxa": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "faturado_produto": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "faturado_servico": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "faturado_total": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "faturado_cmv": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "desp_var": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "desp_fixo": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "margem_contrib": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "margem_contrib_perc": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "lucro_bruto": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "margem_bruta": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "lucro_estimado": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "lucro_pe": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "ticket_medio": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_dinh": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_cheq": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_boleto": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_c_cred": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_c_deb": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_crediario": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_vale": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_depcc": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "tpag_pix": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "estoque_medio": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_incluidos": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_perdidos": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_perdidos_valor": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_faturados": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_faturados_valor": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "dav_taxa_conver": {
                        "description": "",
                        "type": "numeric",
                        "example": "",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data da criacao da empresa",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Data da ultima alteracao da empresa",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}